refactor: extract table views into RadarrTableView and SonarrTableView
